moved functions relates to file copying to module `copy`
